refactor: move fpu i/f to another file
