+++ rtl/main/scala/Twinlight/TLFpu/FPU_ADD.scala
@@ -8,21 +8,6 @@ import scala.language.postfixOps
 
 //============================================       INTERFACES        ============================================//
 
-case class FPU_IF(expWidth: Int, precision: Int) extends Bundle with IMasterSlave {
-  // 1 for the sign bit
-  val inputWidth = expWidth + precision + 1
-
-  val a, b = UInt(inputWidth bits)
-  val rm = RoundingEncoding() // round mode
-  val result = UInt(inputWidth bits)
-  val fflags = UInt(5 bits)
-
-  def asMaster(): Unit = {
-    in(a, b, rm)
-    out(result, fflags)
-  }
-}
-
 case class FPU_ADD_Special_Info() extends Bundle {
   val iv = Bool()
   val nan = Bool()
